requirements.txt
fastapi==0.61.1
uvicorn==0.11.8
pydantic==1.6.1
requests==2.24.0
pycryptodome==3.9.9
python-jenkins==1.7.0
aiofiles==0.6.0
PyJWT==1.7.1
bcrypt==3.2.0
paramiko==2.7.2
sentry-sdk==1.1.0